Move the preference database code into a separate file
